feat: merge delivery + pickup into single /book/[slug] page with toggle
- New /book/[slug] page with Drop-off Delivery / Request Pickup toggle
- Toggle only shows options the buyer supports (has_booking / has_pickup)
- /book-delivery/[slug] and /request-pickup/[slug] now redirect to /book/[slug]
- Buyer profile sidebar consolidated into one card with context-aware description
