+++ apps/client-fnp/components/layouts/client.tsx
@@ -147,35 +147,23 @@ export function Client({ slug, user }: ClientPageProps) {
 
           {/* Right Sidebar - Contact (1/3 width) */}
           <div className="space-y-6">
-            {client.has_booking && client.type === 'buyer' && (
+            {(client.has_booking || client.has_pickup) && client.type === 'buyer' && (
               <div className="bg-blue-50 border border-blue-200 rounded-xl p-6 space-y-3">
                 <div>
                   <p className="font-semibold text-sm text-blue-900">Sell to {capitalizeFirstLetter(client.name)}</p>
                   <p className="text-xs text-blue-700 mt-1 leading-relaxed">
-                    This buyer accepts online delivery bookings. Reserve your drop-off slot directly from the platform.
+                    {client.has_booking && client.has_pickup
+                      ? "Select a drop-off location or request pickup, and pick a date to reserve your slot."
+                      : client.has_booking
+                      ? "Reserve a drop-off slot directly from the platform."
+                      : "Request this buyer to collect goods directly from your farm."}
                   </p>
                 </div>
                 <Link
-                  href={`/book-delivery/${client.name.toLowerCase().replace(/\s+/g, '-')}`}
+                  href={`/book/${client.name.toLowerCase().replace(/\s+/g, '-')}`}
                   className="block w-full text-center bg-blue-700 text-white text-sm font-semibold px-4 py-2.5 rounded-lg hover:bg-blue-800 transition-colors"
                 >
                   Book a Sale →
-                </Link>
-              </div>
-            )}
-            {client.has_pickup && client.type === 'buyer' && (
-              <div className="bg-green-50 border border-green-200 rounded-xl p-6 space-y-3">
-                <div>
-                  <p className="font-semibold text-sm text-green-900">Request Pickup from {capitalizeFirstLetter(client.name)}</p>
-                  <p className="text-xs text-green-700 mt-1 leading-relaxed">
-                    This buyer offers on-farm pickups. Request them to collect goods directly from your farm.
-                  </p>
-                </div>
-                <Link
-                  href={`/request-pickup/${client.name.toLowerCase().replace(/\s+/g, '-')}`}
-                  className="block w-full text-center bg-green-700 text-white text-sm font-semibold px-4 py-2.5 rounded-lg hover:bg-green-800 transition-colors"
-                >
-                  Request Pickup →
                 </Link>
               </div>
             )}
